# api/openapi.json
{
  "components": {
    "schemas": {
...
      "ChartResult": {
        "additionalProperties": false,
        "properties": {
          "captions": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Markdown captions",
            "title": "Captions",
            "type": "object"
          },
          "figures": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "description": "name -> Plotly figure JSON for react-plotly.js",
            "title": "Figures",
            "type": "object"
          },
          "rows": {
            "additionalProperties": {
              "type": "integer"
            },
            "description": "Subplot rows (sizing hint)",
            "title": "Rows",
            "type": "object"
          }
        },
        "required": [
          "figures"
        ],
        "title": "ChartResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Mixing and reactor-engineering calculations, databases, reports and charts for the Mixing Lab front end.",
    "title": "Mixing Lab API",
    "version": "Beta"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/charts/{kind}": {
      "post": {
        "description": "kind: assessment-envelope, assessment-surfaces, comparison-envelope, heat-cool, reaction-profile, ua-surface, bourne-speed-plan, solvent-properties, blend-phases",
        "operationId": "chart_api_v1_charts__kind__post",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Plotly figure JSON for react-plotly.js",
        "tags": [
          "Reports & charts"
        ]
      }
    },
...
    "/api/v1/media/vessels/{name}": {
      "get": {
        "operationId": "vessel_media_api_v1_media_vessels__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Vessel Media Api V1 Media Vessels  Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Best 3D model / image of a vessel",
        "tags": [
          "Reference"
        ]
      }
    },
    "/api/v1/media/vessels/{name}/fill": {
      "get": {
        "operationId": "vessel_fill_api_v1_media_vessels__name__fill_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "fill_L",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fill L"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Vessel Fill Api V1 Media Vessels  Name  Fill Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Liquid level and warnings at a fill volume",
        "tags": [
          "Reference"
        ]
      }
    },
    "/api/v1/media/vessels/{name}/schematic.png": {
      "get": {
        "operationId": "vessel_schematic_api_v1_media_vessels__name__schematic_png_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "fill_L",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fill L"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "image/png": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "2D cross-section drawing",
        "tags": [
          "Reference"
        ]
      }
    },
...
    "/api/v1/reports/{kind}": {
      "post": {
        "description": "kind: assessment, comparison, sensitivity, bourne, heat-cool, reaction-profile",
        "operationId": "report_api_v1_reports__kind__post",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "application/pdf": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build a PDF report",
        "tags": [
          "Reports & charts"
        ]
      }
    },
...
  }
}